// libs/finicity-helpers/src/types/FinicityAccountType.ts
export enum FinicityAccountType {
  Checking = 'checking',
  Savings = 'savings',
  CD = 'cd',
  MoneyMarket = 'moneyMarket',
  CreditCard = 'creditCard',
  LineOfCredit = 'lineOfCredit',
  Investment = 'investment',
  InvestmentTaxDeferred = 'investmentTaxDeferred',
  EmployeeStockPurchasePlan = 'employeeStockPurchasePlan',
  Ira = 'ira',
  _401k = '401k',
  Roth = 'roth',
  _403b = '403b',
  _529plan = '529plan',
  Rollover = 'rollover',
  Ugma = 'ugma',
  Utma = 'utma',
  Keogh = 'keogh',
  _457plan = '457plan',
  _401a = '401a',
  BrokerageAccount = 'brokerageAccount',
  EducationSavings = 'educationSavings',
  HealthSavingsAccount = 'healthSavingsAccount',
  Pension = 'pension',
  ProfitSharingPlan = 'profitSharingPlan',
  Roth401k = 'roth401k',
  SepIRA = 'sepIRA',
  SimpleIRA = 'simpleIRA',
  ThriftSavingsPlan = 'thriftSavingsPlan',
  VariableAnnuity = 'variableAnnuity',
  Cryptocurrency = 'cryptocurrency',
  Mortgage = 'mortgage',
  Loan = 'loan',
  StudentLoan = 'studentLoan',
  StudentLoanGroup = 'studentLoanGroup',
  StudentLoanAccount = 'studentLoanAccount'
}